feat: Create new file for extracted logic
